SCOTT: Initial commit of Return to Pirate's Isle files
SCOTT: Atari 8-bit vector: Reduce fill recursion a bit

common_sagadraw: Make tile rotation routines non-static

AIUKGraphics: irmak.c: Define some magic numbers

SCOTT: Silence analyzer warnings

SCOTT: Create tests for new Scott vector graphics formats

SCOTT: Add support for MAME zip version of RTPI

The same as the .cart version, except each ROM is
a separate file inside the zip.

SCOTT: Add title screen text for RTPI

Tried to make it look nice while keeping
the original intention.

SCOTT: Fix Mysterious Adventures detection

If we have already detected the game, we don't have to do it
the hard way, which may fail.

SCOTT: Fix image height for Mysterious Adventures

There were some black lines at the bottom.

+++ terps/scott/ti994a/ti99_4a_terp.c
@@ -144,7 +144,7 @@ static ActionResultType PerformTI99Line(const uint8_t *action_line)
 #ifdef DEBUG_ACTIONS
             debug_print("Is bitflag %d NOT set?\n", *ptr);
 #endif
-            if (BitFlags & (1 << *(ptr++))) {
+            if (BitFlags & ((uint64_t)1 << (uint64_t)*(ptr++))) {
                 run_code = 1;
                 result = ACT_FAILURE;
             }
